review comments

+++ src/aggregate/aggregate_exec.c
@@ -48,7 +48,7 @@ static int prepareExecutionPlan(AREQ *req, QueryError *status);
 // Try to claim reply ownership. Returns true if claimed (state was NOT_REPLIED),
 // false if already claimed or replied (state was REPLYING or REPLIED).
 static inline bool AREQ_TryClaimReply(AREQ *req) {
-  int expected = ReplyState_NotReplied;
+  uint8_t expected = ReplyState_NotReplied;
   return atomic_compare_exchange_strong_explicit(&req->replyState, &expected,
       ReplyState_Replying, memory_order_acq_rel, memory_order_acquire);
 }
